refactor: :recycle: combine Tle classes and improve documentation
