+++ .eslintignore
@@ -1,3 +1,3 @@
-/dist
 /lib
-/src/sgp4/asc
+/examples/commonjs/satellite-js-migration.js
+/examples/mjs/satellite-js-migration.mjs
