+++ src/enums/Sgp4OpsMode.ts
@@ -1,0 +1,4 @@
+export enum Sgp4OpsMode {
+  AFSPC = 'a',
+  IMPROVED = 'i',
+}
